greyscale image and refactor
